+++ src/app/store.js
@@ -2,6 +2,9 @@ import { configureStore } from '@reduxjs/toolkit';
 import { colorApi } from '../features/api/colors';
 import titleReducer from '../features/creative-title/titleSlice';
 
+// maintaining all state that should be shared across the app
+// and all the reducers that should be used to update that state
+// including the API slice reducer
 const store = configureStore({
   reducer: {
     [colorApi.reducerPath]: colorApi.reducer,
