+++ src/components/pages/home/FilterBy.jsx
@@ -14,6 +14,8 @@ const FilterBy = () => {
   const dispatch = useDispatch();
 
   useEffect(() => {
+    // we can use useCallback here to avoid creating a new function on every render 
+    // but as we are not rerendering (a lot) this component so it's fine to use it here
     const applyFilters = () => {
       const filteredData = filterData(filters, titles);
       dispatch(filteredTitles(filteredData));
