+++ src/style/Home.css
@@ -69,6 +69,7 @@ aside form button {
   border: none;
   border-radius: 5px;
   cursor: pointer;
+  transition: all 0.8s ease;
 }
 
 .background-colors .color-box {
